+++ src/reducers/shortLink.js
@@ -1,6 +1,6 @@
 import { ADD_LINK } from '../constants/AppConstants';
 
-export default function shortLink(state = [], action = {}) {
+export default function shortLink(state = {}, action = {}) {
   switch (action.type) {
     case ADD_LINK:
       return action.link;
